requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0
matplotlib>=3.6.0
seaborn>=0.12.0
geopandas>=0.12.0
h5py>=3.7.0
torch>=2.0.0
pyyaml>=6.0
